requirements.txt
dotenv
streamlit
langchain-gigachat==0.3.13a1
langchain_openai
langchain-core
langgraph
pytest
langchain-tavily
